Fixed compilation warnings/errors under C++ mode
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@1569 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjlib-util/src/pjlib-util-test/encryption.c
@@ -536,7 +536,7 @@ static int base64_test(void)
 	if (rc != PJ_SUCCESS)
 	    return -90;
 
-	if (out_len != strlen(base64_test_vec[i].base64))
+	if (out_len != (int)strlen(base64_test_vec[i].base64))
 	    return -91;
 
 	output[out_len] = '\0';
@@ -551,7 +551,7 @@ static int base64_test(void)
 	if (rc != PJ_SUCCESS)
 	    return -95;
 
-	if (out_len != strlen(base64_test_vec[i].base256))
+	if (out_len != (int)strlen(base64_test_vec[i].base256))
 	    return -96;
 
 	output[out_len] = '\0';
